move integration test boilerplate to a base test class & use funsuite for more concise test naming
